Refactor current test to support multiple provider tests
