+++ src/integrationTest/java/uk/gov/hmcts/reform/bulkscanccdeventhandler/pact/ContractTestSuite.java
@@ -1,18 +1,13 @@
-package uk.gov.hmcts.reform.bulkscanccdeventhandler;
+package uk.gov.hmcts.reform.bulkscanccdeventhandler.pact;
 
-import au.com.dius.pact.provider.junit.Provider;
 import au.com.dius.pact.provider.junit.loader.PactBroker;
 import au.com.dius.pact.provider.junit5.HttpTestTarget;
 import au.com.dius.pact.provider.junit5.PactVerificationContext;
-import au.com.dius.pact.provider.junit5.PactVerificationInvocationContextProvider;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.TestTemplate;
-import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.boot.web.server.LocalServerPort;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
 import uk.gov.hmcts.reform.bulkscanccdeventhandler.common.auth.AuthService;
 
 import static org.mockito.ArgumentMatchers.any;
@@ -20,11 +15,9 @@ import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.when;
 import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.RANDOM_PORT;
 
-@Provider("sample_app_ocr_validation")
-@ExtendWith(SpringExtension.class)
-@SpringBootTest(webEnvironment = RANDOM_PORT)
 @PactBroker(host = "pact-broker.platform.hmcts.net", port = "443", scheme = "https")
-public class OcrValidationContractTest {
+@SpringBootTest(webEnvironment = RANDOM_PORT)
+abstract class ContractTestSuite {
 
     @LocalServerPort
     private int port;
@@ -41,11 +34,5 @@ public class OcrValidationContractTest {
         context.setTarget(new HttpTestTarget("localhost", port, "/"));
         when(authService.authenticate(any())).thenReturn("bulk_scan_sample_app_test");
         doNothing().when(authService).assertIsAllowedService(any());
-    }
-
-    @TestTemplate
-    @ExtendWith(PactVerificationInvocationContextProvider.class)
-    public void pactVerificationTestTemplate(PactVerificationContext context) {
-        context.verifyInteraction();
     }
 }
